Stop. Mutt time.
Mutt setup, based on [Steve Losh's post][slosh], but tailored to my tastes. Bits
stolen from all over the place.

To work correctly, also requires:
* offlineimap
* msmtp
* notmuch
* urlview
* gpg

And a gpg key with the current id. I'll have to fix the hardcoding of that. I've
tried to document the muttrc file well, but I think it needs to be organized by
function, not settings / bindings.

[slosh]: http://stevelosh.com/blog/2012/10/the-homely-mutt/